# Conflicts:
#	fs_brokers/cdc_client/src/test/java/org/apache/doris/cdcclient/utils/ConfigUtilTest.java

+++ be/src/core/value/vdatetime_value.h
@@ -691,9 +691,8 @@ public:
     uint32_t to_date_v2() const { return (year() << 9 | month() << 5 | day()); }
 
     uint64_t to_datetime_v2() const {
-        return (uint64_t)(((uint64_t)year() << 46) | ((uint64_t)month() << 42) |
-                          ((uint64_t)day() << 37) | ((uint64_t)hour() << 32) |
-                          ((uint64_t)minute() << 26) | ((uint64_t)second() << 20));
+        return (((uint64_t)year() << 46) | ((uint64_t)month() << 42) | ((uint64_t)day() << 37) |
+                ((uint64_t)hour() << 32) | ((uint64_t)minute() << 26) | ((uint64_t)second() << 20));
     }
 
     uint32_t hash(int seed) const { return HashUtil::hash(this, sizeof(*this), seed); }
